% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-param.R
\name{extract_param}
\alias{extract_param}
\alias{extract_theta}
\alias{extract_omega}
\alias{extract_sigma}
\title{Extract initial estimates from parameter records}
\usage{
extract_theta(records, mark_flags = NULL)

extract_omega(records, mark_flags = NULL)

extract_sigma(records, mark_flags = NULL)
}
\arguments{
\item{records}{An \link{nmrec_ctl_records} object.}

\item{mark_flags}{A vector of NONMEM flags (i.e. valueless options such as
\code{FIXED} or \code{SD}). For each specified flag, construct a boolean vector (for
\code{THETA}) or matrix (for \code{OMEGA} and \code{SIGMA}) indicating whether the flag is
"active" for the value. Any valid spelling of the flag name is allowed.

For example, passing a value of "fix" would indicate whether a \code{FIXED} flag
is in effect for each value. The flag may be linked to an individual
initial estimate (e.g., \code{FIXED} for a \code{THETA} or diagonal \code{OMEGA}record) or
linked to all values in the record (e.g., \code{FIXED} for a \code{THETA} or block
\code{OMEGA} record). In either case, the boolean vector or matrix is always the
same shape as the return value.

The results are stored in a named list as the "nmrec_flags" attribute
attached to the return value, with the names corresponding to the resolved
flag names.}
}
\value{
A vector (for \code{THETA}) or a square matrix (for \code{OMEGA} and \code{SIGMA}).
For matrix values, the upper triangle is always filled with \code{NA} values.
}
\description{
Construct a numeric object from initial estimates defined by control stream
records:
\itemize{
\item \code{extract_theta()} returns a vector of initial estimates defined by \code{THETA}
records.
\item \code{extract_omega()} and \code{extract_sigma()} return a matrix of initial
estimates for \code{OMEGA} and \code{SIGMA} records.
}
}
\details{
\subsection{Constraints and limitations}{
\itemize{
\item These functions return initial estimates only if they are \strong{explicitly}
defined in the control stream. All other values are returned as \code{NA}. For
example, a record of \verb{$THETA (1)x4} is returned as \code{c(1, NA, NA, NA)}.
\item If additional parameter records are used for priors, those are included in
the result. To avoid this, instead use informative prior record names
(such as \code{THETAP} and \code{THETAPV}).
}
}
}
\examples{
ctl <- parse_ctl(c(
  "$PROBLEM ex",
  "$THETA 2 FIX",
  "$THETA (10,30)x2",
  "$OMEGA BLOCK(3)",
  "0.1",
  "0.01 0.1",
  "0.01 0.01 0.1"
))
extract_theta(ctl)
extract_theta(ctl, mark_flags = "fixed")

extract_omega(ctl)
}
\seealso{
\link{set_param} for setting parameter options from values
}
