% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-record-option.R
\name{get_record_option}
\alias{get_record_option}
\title{Extract an option from a record}
\usage{
get_record_option(record, name)
}
\arguments{
\item{record}{An \link{nmrec_record} object.}

\item{name}{Name of option to select. Any valid spelling of the option name
is allowed.}
}
\value{
An \link{nmrec_option} object. \code{NULL} is returned if an option for \code{name}
is not found in \code{record}.
}
\description{
A parsed \link{nmrec_record} object has a populated "values" field that includes
\link{nmrec_option} objects. This function is a convenience wrapper for grabbing a
specified option. If the option occurs multiple times in the record, an error
is signaled.
}
\details{
Note that a record also stores the record type (e.g., "$EST") as an
\link{nmrec_option} object (specifically an \code{nmrec_option_record_name} object).
This "option" won't usually be of interest, but, if you do want to grab it
with \code{get_record_option()}, you need to pass the full type as \code{name} (e.g.,
"estimation" rather than "est").
}
\examples{
ctl <- parse_ctl(nmrec_examples[["bayes1"]])
ests <- select_records(ctl, "est")
# Get method for first estimation record.
meth1 <- get_record_option(ests[[1]], "meth")
meth1$value

}
\seealso{
\code{\link[=select_records]{select_records()}} selecting records of a given type.
}
