% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrec.R
\docType{package}
\name{nmrec-package}
\alias{nmrec}
\alias{nmrec-package}
\title{nmrec: Read, Parse, and Modify NONMEM Control Records}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An interface reading records from a NONMEM control stream, parsing and modifying records of interest, and writing the result.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://metrumresearchgroup.github.io/nmrec}
  \item \url{https://github.com/metrumresearchgroup/nmrec}
  \item Report bugs at \url{https://github.com/metrumresearchgroup/nmrec/issues}
}

}
\author{
\strong{Maintainer}: Kyle Meyer \email{kylem@metrumrg.com}

Authors:
\itemize{
  \item Kyle Barrett \email{barrettk@metrumrg.com}
}

}
\keyword{internal}
