% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ctl.R
\name{nmrec_ctl_records}
\alias{nmrec_ctl_records}
\alias{ctl_records}
\title{NONMEM control records object}
\description{
S3 object representing a NONMEM control stream.  This has two fields:
\itemize{
\item \code{frontmatter}: a character vector of lines, if any, before the first
record
\item \code{records}: a list of \link{record} objects
}
}
