% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{nmrec_examples}
\alias{nmrec_examples}
\title{Environment of containing control streams examples}
\format{
An object of class \code{environment} of length 4.
}
\usage{
nmrec_examples
}
\description{
These are selected from NONMEM 7.5 examples. Some comments and blank lines
have been removed.
}
\examples{

# Get list of available examples.
ls(nmrec_examples)

# Parse an example into an `nmrec_ctl_records` object.
ctl <- parse_ctl(get("bayes1", envir = nmrec_examples))
head(ctl$records)

}
\keyword{datasets}
