% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{option}
\alias{option}
\alias{nmrec_option}
\alias{option_pos}
\alias{option_flag}
\alias{option_value}
\alias{option_record_name}
\alias{option_nested}
\title{Option objects}
\description{
Parsing a \link{record} populates its \code{values} field with elements and options. An
option is an \link[R6:R6Class]{R6} object that inherit from \code{nmrec_option}.
}
\details{
\subsection{Types of options}{
\itemize{
\item \strong{positional} (\code{nmrec_option_pos}): the option is rendered as the \code{value}
field.

For example, the file name from a \verb{$DATA} record with a value of
"foo.csv" would be rendered as "foo.csv".
\item \strong{flag} (\code{nmrec_option_flag}): if \code{value} is not \code{NULL}, the option is
rendered as the \code{name_raw} field.

For example, the interaction flag for an \verb{$ESTIMATION} record with a
\code{name_raw} value of "interact" and a value of \code{TRUE} would be rendered as
"interact".
\item \strong{value} (\code{nmrec_option_value}): render name-value pair from \code{name_raw}
field and \code{value} field, separating the values by the \code{sep} field value.

For example, the maxevals option for a \verb{$ESTIMATION} record with a
\code{name_raw} field of "MAX", a \code{value} field of "999", and \code{sep} field of
" = " would be rendered as "MAX = 999".
\item \strong{record name} (\code{nmrec_option_record_name}): this option derives from
\code{nmrec_option_flag} and tailors formatting for record names.

For example, a \code{name_raw} value of "est" for an \verb{$ESTIMATION} record would
be rendered as "$est".
\item \strong{nested} (\code{nmrec_option_nested}): this option derives from
\code{nmrec_option_pos}, replacing the simpler \code{value} field with a \code{values}
field that is a list of items to render. Like the \code{values} field of a
\link{record} objects, this consists of elements and options. It is useful for
\code{THETA}, \code{OMEGA}, and \code{SIGMA} values, which can have more complex
formatting and have attached options (such as "FIXED").
}
}
}
\seealso{
\code{\link[=get_record_option]{get_record_option()}} for getting a specified option from a \link{record}
object.
}
