% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ctl.R
\name{read_ctl}
\alias{read_ctl}
\alias{parse_ctl}
\title{Read records from a NONMEM control stream}
\usage{
read_ctl(path)

parse_ctl(lines)
}
\arguments{
\item{path}{Path to NONMEM control stream.}

\item{lines}{Lines from a NONMEM control stream.}
}
\value{
A \code{nmrec_ctl_records} object.
}
\description{
\code{read_ctl()} parses records from a file, and \code{parse_ctl()} parses them from a
character vector.
}
