% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record-data.R, R/record-estimation.R,
%   R/record-input.R, R/record-omega.R, R/record-prior.R, R/record-problem.R,
%   R/record-sigma.R, R/record-table.R, R/record-theta.R, R/record.R
\name{record_data}
\alias{record_data}
\alias{record_estimation}
\alias{record_input}
\alias{record_omega}
\alias{record_prior}
\alias{record_problem}
\alias{record_sigma}
\alias{record_table}
\alias{record_theta}
\alias{record}
\alias{nmrec_record}
\alias{record_raw}
\title{Record objects}
\description{
When \code{\link[=read_ctl]{read_ctl()}} reads a control stream, it returns an \link{nmrec_ctl_records}
S3 object. Its \code{records} field is a list of \link[R6:R6Class]{R6} objects that
inherits from \code{nmrec_record}. When parsing \emph{within} a record is supported for
a given record type (e.g., "estimation"), the child class is specific for
that type (e.g., \code{nmrec_record_estimation}). On the other hand, records for
which parsing has not been implemented have a type \code{nmrec_record_raw}.
}
\details{
\subsection{Fields}{
\itemize{
\item \code{name}: the standardized name for the record (e.g., "estimation").
\item \code{values}: a list capturing the record's content. This is constructed by
the \verb{$parse()} method and consists of "elements" like white space and
line breaks interspersed with \link{option} objects.
}
}

\subsection{Methods}{
\itemize{
\item \code{parse}: parse the record's lines into \code{values}, enabling inspection and
modification.
\item \code{get_options}: return list of \link{nmrec_option} objects from \code{values}.
\item \code{get_lines}: return original \code{lines} passed during initialization. These
values are never updated.
\item \code{format}: render the record as a string. If the record has been parsed,
this is derived from \code{values}.
}
}
}
\seealso{
\link{select_records} to get records of a given type from an
\link{nmrec_ctl_records} object.
}
