% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-records.R
\name{select_records}
\alias{select_records}
\title{Select records of the specified type}
\usage{
select_records(records, name)
}
\arguments{
\item{records}{An \link{nmrec_ctl_records} object.}

\item{name}{Get \link{record} objects of this type. This may be spelled any way
that's accepted in the control stream.}
}
\value{
A list of \link{nmrec_record} objects.
}
\description{
An \link{nmrec_ctl_records} object has all the records of a control stream under
its "records" field. This function is a convenience wrapper for selecting
records of a given type (e.g., "theta").
}
\examples{
ctl <- parse_ctl(nmrec_examples[["control3"]])
# Get all "$ESTIMATION" records.
select_records(ctl, "est")

}
\seealso{
\code{\link[=read_ctl]{read_ctl()}} for reading records from a file, \code{\link[=get_record_option]{get_record_option()}}
for extracting an option from a record.
}
