% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-param.R
\name{set_param}
\alias{set_param}
\alias{set_theta}
\alias{set_omega}
\alias{set_sigma}
\title{Set initial estimates in parameter records}
\usage{
set_theta(records, values, fmt = "\%.3G", bounds = c("keep", "discard"))

set_omega(records, values, fmt = "\%.3G", representation = c("keep", "reset"))

set_sigma(records, values, fmt = "\%.3G", representation = c("keep", "reset"))
}
\arguments{
\item{records}{An \link{nmrec_ctl_records} object.}

\item{values}{Overwrite the parameter's initial estimates in \code{records} with
these values. The length must match the full length defined by the
combination of the parameter's records.

For \code{THETA}, \code{values} must be a vector. For \code{OMEGA} and \code{SIGMA}, \code{values} can be
\enumerate{
\item the square matrix, in which case the values in the lower triangle
(including the diagonal) are used
\item a vector in row-major order for values in the lower triangle (including
diagonal).
}

Use \code{NA} in \code{values} to skip updating the corresponding option.}

\item{fmt}{Convert each value to a string with this \code{\link[=sprintf]{sprintf()}} format
specifier.}

\item{bounds}{Whether to keep or discard the existing bounds when setting the
initial estimates in \code{THETA} records.}

\item{representation}{Whether to keep alternative representation options like
\code{SD} and \code{CORRELATION} or reset to the default representation (variance and
covariance).

For \code{OMEGA} and \code{SIGMA} records, NONMEM supports specifying diagonal and
off-diagonal initial estimates in a different representation than the
default, variance and covariance. If \code{values} are the final estimates from
a previous NONMEM run, the alternative representation options should be
discarded because NONMEM always outputs variances and covariances.

\code{NA} values can lead to records and options being "untouched" (i.e. not
overwritten with the value from \code{values}). Even when "reset" is specified,
alternative options are not discarded for untouched \emph{records} (in the case
of BLOCK records, where the option applies to all values in the block) or
for untouched \emph{options} (in the case of \code{DIAGONAL} records, where the
options apply to individual initial estimates).}
}
\description{
\code{set_theta()}, \code{set_omega()}, and \code{set_sigma()} enable updating \code{THETA},
\code{OMEGA}, and \code{SIGMA} records by specifying a full-length parameter vector or
matrix. These functions handle mapping from the parameter space to the
corresponding option.
}
\details{
\subsection{Constraints and limitations}{
\itemize{
\item These functions update initial estimates only if they are \strong{explicitly}
defined in the control stream.

For example, consider the update of \verb{$THETA (1)x4}. That defines four
initial estimates, but only the first explicitly appears. Calling
\code{set_theta()} with a value of \code{c(5, 6, 7, 8)} gives a result of \verb{(5)x4}.
\item The caller must specify a value with a size that matches what is defined
by the parameter records.
\item Using additional parameter records for priors is not supported and will
lead to a size mismatch between the parameter and its records. Instead use
informative prior record names (such as \code{THETAP} and \code{THETAPV}).
}
}
}
\examples{
ctl <- parse_ctl(c(
  "$PROBLEM ex",
  "$THETA 2",
  "$THETA (30)x2",
  "$OMEGA BLOCK(3)",
  "0.1",
  "0.01 0.1",
  "0.01 0.01 0.1"
))
set_theta(ctl, c(10, 40, 20))
ctl

new_omega <- matrix(
  c(
    0.5, 0.2, 0.2,
    0.2, NA, 0.2,
    0.2, 0.2, 0.7
  ),
  nrow = 3, byrow = TRUE
)
# The upper triangle doesn't come into play; let's zero it for clarity, but
# it would be fine to leave as is.
new_omega[upper.tri(new_omega)] <- 0
new_omega
set_omega(ctl, new_omega)
ctl
}
\seealso{
\code{\link[=set_record_option]{set_record_option()}} setting option by name.
}
