% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-record-option.R
\name{set_record_option}
\alias{set_record_option}
\title{Add or modify a record option}
\usage{
set_record_option(record, name, value)
}
\arguments{
\item{record}{An \link{nmrec_record} object.}

\item{name}{Name of option to select. Any valid spelling of the option name
is allowed.}

\item{value}{Value to assign to option. For a value option, the specified
value can be anything that can be formatted to a string. For a flag option, any
non-\code{NULL} value results in the name being included in the record. Passing
\code{NULL} drops the option from the displayed record.}
}
\description{
Add a new flag or value option or adjust the value/presence of an existing one.
}
\examples{
ctl <- parse_ctl(nmrec_examples[["bayes1"]])
ests <- select_records(ctl, "est")
# Get method for first estimation record.
set_record_option(ests[[1]], "meth", "nuts")
ests[[1]]

}
\seealso{
\link{nmrec_option} for a description of option types
}
