% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-ctl.R
\name{write_ctl}
\alias{write_ctl}
\title{Write records to a file}
\usage{
write_ctl(records, path)
}
\arguments{
\item{records}{A \code{nmrec_ctl_records} object.}

\item{path}{File to write records to.}
}
\value{
\code{records}, unmodified and invisibly.
}
\description{
Generate a new control stream from a \code{nmrec_ctl_records} object.
}
\seealso{
\code{\link[=read_ctl]{read_ctl()}} for reading records from a file.
}
