% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrec.R
\docType{package}
\name{nmrec-package}
\alias{nmrec}
\alias{nmrec-package}
\title{nmrec: Read, Parse, and Modify NONMEM control records}
\description{
A library for reading records from a NONMEM control stream, parsing and modifying records of interest, and writing the result.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/metrumresearchgroup/nmrec}
  \item Report bugs at \url{https://github.com/metrumresearchgroup/nmrec/issues}
}

}
\author{
\strong{Maintainer}: Kyle Meyer \email{kylem@metrumrg.com}

}
\keyword{internal}
