% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-param.R
\name{set_param}
\alias{set_param}
\alias{set_theta}
\alias{set_omega}
\alias{set_sigma}
\title{Set initial estimates in parameter records}
\usage{
set_theta(records, values)

set_omega(records, values)

set_sigma(records, values)
}
\arguments{
\item{records}{An \link{nmrec_ctl_records} object.}

\item{values}{Overwrite the parameter's initial estimates in \code{records} with
these values. The length must match the full length defined by the
combination of the parameter's records.

For THETA, \code{values} must be a vector. For OMEGA and SIGMA, \code{values} can be
\enumerate{
\item the square matrix, in which case the values in the lower triangle
(including the diagonal) are used
\item a vector in row-major order for values in the lower triangle (including
diagonal).
}

Use \code{NA} in \code{values} to skip updating the corresponding option.}
}
\description{
\code{set_theta()}, \code{set_omega()}, and \code{set_sigma()} enable updating THETA, OMEGA,
and SIGMA records by specifying a full-length parameter vector or matrix.
These functions handle mapping from the parameter space to the corresponding
option.
}
\details{
\subsection{Constraints and limitations}{
\itemize{
\item These functions update initial estimates only if they are \strong{explicitly}
defined in the control stream.

For example, consider the update of \verb{$THETA (1)x4}. That defines four
initial estimates, but only the first explicitly appears. Calling
\code{set_theta()} with a value of \code{c(5, 6, 7, 8)} gives a result of \verb{(5)x4}.
\item The caller must specify a value with a size that matches what is defined
by the parameter records.
\item Using additional parameter records for priors is not supported and will
lead to a size mismatch between the parameter and its records. Instead use
informative prior record names (such as THETAP and THETAPV).
}
}
}
\examples{
ctl <- parse_ctl(c(
  "$PROBLEM ex",
  "$THETA 2",
  "$THETA (30)x2",
  "$OMEGA BLOCK(3)",
  "0.1",
  "0.01 0.1",
  "0.01 0.01 0.1"
))
set_theta(ctl, c(10, 40, 20))
ctl

new_omega <- matrix(
  c(
    0.5, 0.2, 0.2,
    0.2, NA, 0.2,
    0.2, 0.2, 0.7
  ),
  nrow = 3, byrow = TRUE
)
# The upper triangle doesn't come into play; let's zero it for clarity, but
# it would be fine to leave as is.
new_omega[upper.tri(new_omega)] <- 0
new_omega
set_omega(ctl, new_omega)
ctl
}
\seealso{
\code{\link[=set_record_option]{set_record_option()}} setting option by name.
}
